Extract scheduler into its own module
The scheduler is getting quite complicated. I'll extract it into
its own module.
